requirements/dev-requirements.txt
-r optional_requirements.txt

pytest
gempy
imagehash
pillow
pydantic
gempy_engine>=2025.2.0dev0,<2025.3.0
gempy>=2025.2.0dev0,<2025.3.0
